+++ src/main/java/at/sv/hue/api/PutCall.java
@@ -3,6 +3,9 @@ package at.sv.hue.api;
 import lombok.Builder;
 import lombok.Data;
 
+import java.util.Objects;
+import java.util.stream.Stream;
+
 @Data
 @Builder(toBuilder = true)
 public final class PutCall {
@@ -17,4 +20,8 @@ public final class PutCall {
     String effect;
     Integer transitionTime;
     boolean groupState;
+    
+    public boolean isNullCall() {
+        return Stream.of(bri, ct, x, y, hue, sat, on, effect, transitionTime).allMatch(Objects::isNull);
+    }
 }